requirements.txt
django==4.2.1
channels[daphne]==4.0.0
django-cors-headers==4.0.0
pyyaml
requests==2.31.0
psycopg2-binary